+++ src/services/ShippingMethods.php
@@ -138,10 +138,12 @@ class ShippingMethods extends Component
 
         /** @var ShippingMethod $method */
         foreach ($event->getShippingMethods() as $method) {
+            $totalPrice = $method->getPriceForOrder($order);
+
             if ($method->getIsEnabled() && $method->matchOrder($order)) {
                 $matchingMethods[$method->getHandle()] = [
                     'method' => $method,
-                    'price' => $method->getPriceForOrder($order), // Store the price so we can sort on it before returning
+                    'price' => $totalPrice, // Store the price so we can sort on it before returning
                 ];
             }
         }